Moved event handling logic out of the muzzle.controller and into a event.controller
